requirements.txt
amqp==2.6.1
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==2.2.5
importlib-metadata==4.8.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
prompt-toolkit==3.0.39
protobuf==3.20.3
pytz==2023.3.post1
redis==5.0.1
typing-extensions==4.7.1
vine==1.3.0
wcwidth==0.2.9
Werkzeug==2.2.3
zipp==3.15.0
